Move logger and export exclude to configuration.jl
